+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- site script starts anvil instead of using start
 - Site script works with wallet<>handler separation
 - have site script pass in subtree updater address to deploy script
 - site script works with deposit screener